+++ ContentPatcher/Framework/Patches/EditMapPatch.cs
@@ -46,8 +46,6 @@ namespace ContentPatcher.Framework.Patches
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="path">The path to the patch from the root content file.</param>
-        /// <param name="parentPatch">The parent patch for which this patch was loaded, if any.</param>
-        /// <param name="contentPack">The content pack which requested the patch.</param>
         /// <param name="assetName">The normalized asset name to intercept.</param>
         /// <param name="conditions">The conditions which determine whether this patch should be applied.</param>
         /// <param name="fromAsset">The asset key to load from the content pack instead.</param>
@@ -55,18 +53,22 @@ namespace ContentPatcher.Framework.Patches
         /// <param name="toArea">The map area to overwrite.</param>
         /// <param name="mapProperties">The map properties to change when editing a map, if any.</param>
         /// <param name="mapTiles">The map tiles to change when editing a map.</param>
+        /// <param name="updateRate">When the patch should be updated.</param>
+        /// <param name="contentPack">The content pack which requested the patch.</param>
+        /// <param name="parentPatch">The parent patch for which this patch was loaded, if any.</param>
         /// <param name="monitor">Encapsulates monitoring and logging.</param>
         /// <param name="normalizeAssetName">Normalize an asset name.</param>
-        public EditMapPatch(LogPathBuilder path, IPatch parentPatch, ManagedContentPack contentPack, IManagedTokenString assetName, IEnumerable<Condition> conditions, IManagedTokenString fromAsset, TokenRectangle fromArea, TokenRectangle toArea, IEnumerable<EditMapPatchProperty> mapProperties, IEnumerable<EditMapPatchTile> mapTiles, IMonitor monitor, Func<string, string> normalizeAssetName)
+        public EditMapPatch(LogPathBuilder path, IManagedTokenString assetName, IEnumerable<Condition> conditions, IManagedTokenString fromAsset, TokenRectangle fromArea, TokenRectangle toArea, IEnumerable<EditMapPatchProperty> mapProperties, IEnumerable<EditMapPatchTile> mapTiles, UpdateRate updateRate, ManagedContentPack contentPack, IPatch parentPatch, IMonitor monitor, Func<string, string> normalizeAssetName)
             : base(
                 path: path,
                 type: PatchType.EditMap,
-                contentPack: contentPack,
                 assetName: assetName,
+                fromAsset: fromAsset,
                 conditions: conditions,
-                normalizeAssetName: normalizeAssetName,
+                updateRate: updateRate,
+                contentPack: contentPack,
                 parentPatch: parentPatch,
-                fromAsset: fromAsset
+                normalizeAssetName: normalizeAssetName
             )
         {
             this.FromArea = fromArea;